requirements.txt
pandas
numpy==2.1
seaborn
matplotlib
scipy
scikit-learn
umap-learn
lightgbm
torch
xgboost
sklearn-lvq
memory_profiler